+++ src/components/group/InviteGuestList.tsx
@@ -1,50 +1,49 @@
-import { useDeleteInviteForm } from "@/api/exercise/InviteGuestApi";
+import InviteGuestItem from "@/components/group/InviteGuestItem";
+import Female from "@/assets/icons/female.svg?react";
+import Male from "@/assets/icons/male.svg?react";
+import { useInviteGuest } from "@/api/exercise/InviteGuestApi";
 import type { ResponseInviteGuest } from "@/types/guest";
-import { userLevelMapper } from "@/utils/levelValueExchange";
-import { Member } from "@/components/common/contentcard/Member";
-const { toKor } = userLevelMapper();
 
-type InviteGuestProps = {
-  data?: { list: ResponseInviteGuest[] };
+interface InviteGuestListProps {
   exerciseId: number;
-};
-
-export default function InviteGuestList({
-  data,
-  exerciseId,
-}: InviteGuestProps) {
-  const handleDelete = useDeleteInviteForm(exerciseId);
-
-  //운동 상세 조회
-
-  return data?.list.map((item: ResponseInviteGuest) => {
-    const apilevel = toKor(item.level);
-    console.log(data);
-    const responseLevelValue =
-      apilevel === "disabled"
-        ? "급수 없음"
-        : ["A조", "B조", "C조", "D조"].includes(apilevel)
-          ? `전국 ${apilevel}`
-          : apilevel;
-    const numberStatus = item.isWaiting ? "waiting" : "Participating";
-    const apiNumber = item.isWaiting
-      ? `대기.${item.participantNumber}`
-      : `No.${item.participantNumber}`;
-    return (
-      <Member
-        key={item.guestId}
-        status={numberStatus}
-        {...item}
-        guestName={item.inviterName}
-        gender={item.gender.toUpperCase() as "MALE" | "FEMALE"}
-        number={apiNumber}
-        level={responseLevelValue}
-        showDeleteButton={true}
-        useDeleteModal={false}
-        isGuest={true}
-        guestNumber={true}
-        onDelete={() => handleDelete.mutate(item.guestId)}
-      />
-    );
-  });
 }
+
+export const InviteGuesetList = ({ exerciseId }: InviteGuestListProps) => {
+  const { data } = useInviteGuest(exerciseId);
+
+  const noneData = data.list.length === 0;
+  return (
+    <section>
+      {noneData ? (
+        <div>게스트를 초대해주세요!</div>
+      ) : (
+        <div className="flex flex-col gap-2">
+          <div className="flex items-center justify-between">
+            <div className="flex items-center gap-2">
+              <label className="text-left header-h5">초대된 인원</label>
+              <p className="header-h5">{data.totalCount}</p>
+            </div>
+            <div className="flex items-center gap-2">
+              <Female className="w-4 h-4" />
+              <p className="body-rg-500">{data.femaleCount}</p>
+              <Male className="w-4 h-4" />
+              <p className="body-rg-500">{data.maleCount}</p>
+            </div>
+          </div>
+        </div>
+      )}
+
+      <div className="flex items-center justify-center flex-col">
+        <div className="flex flex-col gap-2">
+          {data.list.map((item: ResponseInviteGuest) => (
+            <InviteGuestItem
+              key={item.guestId}
+              item={item}
+              exerciseId={exerciseId}
+            />
+          ))}
+        </div>
+      </div>
+    </section>
+  );
+};
